backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
easyocr==1.7.1
python-multipart==0.0.6
